drivers/serializers.py
from rest_framework import serializers

from users.serializers import UserSerializer

from .models import *


class DriverSerializer(serializers.ModelSerializer):
    class Meta:
        model = Driver
        fields = '__all__'


class DriverGetSerializer(serializers.ModelSerializer):
    user = UserSerializer()

    class Meta:
        model = Driver
        fields = '__all__'
